Use queryClient as singleton
