+++ client/src/lib/queryClient.ts
@@ -2,8 +2,41 @@ import router from "@shared/contracts";
 import { initQueryClient } from "@ts-rest/react-query";
 
 import { env } from "@/env";
+import { QueryClient } from "@tanstack/react-query";
 
-export const client = initQueryClient(router, {
+const client = initQueryClient(router, {
   validateResponse: true,
   baseUrl: env.NEXT_PUBLIC_API_URL,
 });
+
+function makeQueryClient() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        // With SSR, we usually want to set some default staleTime
+        // above 0 to avoid refetching immediately on the client
+        staleTime: 60 * 1000,
+        refetchOnReconnect: false,
+        refetchOnWindowFocus: false,
+      },
+    },
+  });
+}
+
+let browserQueryClient: QueryClient | undefined = undefined;
+
+function getQueryClient() {
+  if (typeof window === "undefined") {
+    // Server: always make a new query client
+    return makeQueryClient();
+  } else {
+    // Browser: make a new query client if we don't already have one
+    // This is very important so we don't re-make a new client if React
+    // suspends during the initial render. This may not be needed if we
+    // have a suspense boundary BELOW the creation of the query client
+    if (!browserQueryClient) browserQueryClient = makeQueryClient();
+    return browserQueryClient;
+  }
+}
+
+export { client, getQueryClient };
